Changes:  1) added setup.py to make the project installable  2) changed compatibility from py2 to p3  3) switched path resolution from os.path to pathlib for better cross-platform support  4) miscellaneous updates in the code related to 1) or 2)
