+++ README.md
@@ -1,15 +1,13 @@
 # PokéPython
 Your pokemon player is randomly selected from a total of 15 pokemon sprites every time you restart or advance a level. There are 4 levels of progressively more difficult questions on the Python language, and each correct answer will give a speed boost of 100m (in a 1000m dash). All questions are taken from [My Python Quiz](http://mypythonquiz.com/).
 
-Run `python game.py` to play.
-
-Modules: Tkinter, tkFont, pygame
+Install with `pip install .` from the project root, and run with `pokepython`.
 
 ![Playing](/../screenshots/screenshot1.png "Playing")
 ![Wrong Answer](/../screenshots/screenshot2.png "Wrong Answer")
 ![Win](/../screenshots/screenshot3.png "Win")
 
-Music: Clean Bandit - Rather Be feat. Jess Glynne (Instrumental)
+Music: Clean Bandit - Rather Be feat. Jess Glynne (8-bit version)
 
 GIFs: [cloudylicious](http://cloudylicious.deviantart.com/) and [chi-u](chi-u.deviantart.com) on DeviantArt
